Added hash_table.py and MOCK_DATA.csv
Commit One(1) of my hash table homework. Uses a linear probing insert method with the size of the tables being the same as the size of the csv dataset. There were no empty spaces left between the two hash tables, however, the time taken for both tables to be created was about 15 seconds. In addition, the amount of collisions between the two hash tables was quite large, with 91550138 collisions from the title-oriented hash table, and 96570520 collisions from the  quote-oriented hash table. I suspect the hash-key calculating function is what's causing this amount of collisions, and I plan on improving it in my next commit. I also hope that this commit will be my least optimized version of these hash tables, so that I may see heavy improvements when i get to my fifth and final commit.