Updated hash_table.py
Commit two(2) of my hash table homework. The program's insertion method is the same as my last commit, so the amount of empty spaces is still zero, but the one thing I changed was how I calculated my hash key, my function for it's creation now using bit-shifting and a for-loop. The change in my hash key function improved my program drastically, as the hash tables now take around half a second to be created. As well as this, the amount of collisions that occurred dropped significantly, as the title-oriented hash table encountered 854650 collisions, while the quote-oriented hash table encountered 1302967 collisions. I am happy with how much better this second commit has performed compared to the first.